# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Libyan E-Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: List courses with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: language
          in: query
          schema:
            type: string
            enum: [ar, en]
        - name: teacherId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoursesResponse"

    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (teacher only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update course (teacher only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course (teacher only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses/{courseId}/lessons:
    get:
      operationId: getCourseLessons
      tags: [lessons]
      summary: Get lessons for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"

    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson in a course (teacher only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /courses/{courseId}/lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson details (enrolled students/teacher only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"

    put:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson (teacher only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonRequest"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson (teacher only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses/{courseId}/enroll:
    post:
      operationId: enrollCourse
      tags: [enrollments]
      summary: Enroll in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Enrolled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /courses/{courseId}/reviews:
    get:
      operationId: getCourseReviews
      tags: [reviews]
      summary: Get reviews for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
          nullable: true
        courseCount:
          type: integer
      required:
        - id
        - name
        - nameAr
        - courseCount
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
        currency:
          type: string
          default: LYD
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        language:
          type: string
          enum: [ar, en]
        categoryId:
          type: integer
        teacherId:
          type: integer
        teacherName:
          type: string
        teacherAvatar:
          type: string
          nullable: true
        rating:
          type: number
        reviewCount:
          type: integer
        enrollmentCount:
          type: integer
        lessonCount:
          type: integer
        totalDuration:
          type: integer
          description: Total duration in minutes
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - titleAr
        - description
        - descriptionAr
        - price
        - currency
        - level
        - language
        - categoryId
        - teacherId
        - teacherName
        - rating
        - reviewCount
        - enrollmentCount
        - lessonCount
        - totalDuration
        - isPublished
        - createdAt
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/LessonSummary"
            category:
              $ref: "#/components/schemas/Category"
            teacher:
              $ref: "#/components/schemas/TeacherProfile"
            isEnrolled:
              type: boolean
            userProgress:
              type: number
              nullable: true
          required:
            - lessons
            - isEnrolled
    CoursesResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - courses
        - total
        - page
        - limit
        - totalPages
...
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        language:
          type: string
          enum: [ar, en]
        categoryId:
          type: integer
      required:
        - title
        - titleAr
        - description
        - descriptionAr
        - price
        - level
        - language
        - categoryId
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        language:
          type: string
          enum: [ar, en]
        categoryId:
          type: integer
        isPublished:
          type: boolean
    LessonSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        order:
          type: integer
        isFree:
          type: boolean
        type:
          type: string
          enum: [video, text, quiz]
      required:
        - id
        - title
        - titleAr
        - duration
        - order
        - isFree
        - type
    Lesson:
      allOf:
        - $ref: "#/components/schemas/LessonSummary"
        - type: object
          properties:
            courseId:
              type: integer
            videoUrl:
              type: string
              nullable: true
            content:
              type: string
              nullable: true
            contentAr:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
          required:
            - courseId
            - createdAt
    LessonDetail:
      allOf:
        - $ref: "#/components/schemas/Lesson"
        - type: object
          properties:
            isCompleted:
              type: boolean
            watchedSeconds:
              type: integer
          required:
            - isCompleted
            - watchedSeconds
    CreateLessonRequest:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        videoUrl:
          type: string
        content:
          type: string
        contentAr:
          type: string
        duration:
          type: integer
        order:
          type: integer
        isFree:
          type: boolean
        type:
          type: string
          enum: [video, text, quiz]
      required:
        - title
        - titleAr
        - duration
        - order
        - type
    UpdateLessonRequest:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        videoUrl:
          type: string
        content:
          type: string
        contentAr:
          type: string
        duration:
          type: integer
        order:
          type: integer
        isFree:
          type: boolean
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        userId:
          type: integer
        enrolledAt:
          type: string
          format: date-time
        progress:
          type: number
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - courseId
        - userId
        - enrolledAt
        - progress
...
    Review:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - courseId
        - userId
        - userName
        - rating
        - createdAt
    CreateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - rating
...
    TeacherProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        fullNameAr:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        bioAr:
          type: string
          nullable: true
        courseCount:
          type: integer
        studentCount:
          type: integer
        rating:
          type: number
        expertise:
          type: string
          nullable: true
      required:
        - id
        - fullName
        - fullNameAr
        - courseCount
        - studentCount
        - rating